Move common gRPC server code.
Move the common gRPC server initialization code into a helper function.
