+++ pkg/cli/defaults.go
@@ -13,7 +13,35 @@
 
 package cli
 
-import "github.com/spf13/cobra"
+import (
+	"fmt"
+	"os"
+
+	"github.com/projectcontour/contour-authserver/pkg/auth"
+	"github.com/projectcontour/contour-authserver/pkg/version"
+
+	"github.com/spf13/cobra"
+	"google.golang.org/grpc"
+)
+
+func mustString(s string, err error) string {
+	if err != nil {
+		fmt.Fprintf(os.Stderr, "%s: %s\n", version.Progname, err)
+		os.Exit(int(EX_CONFIG))
+	}
+
+	return s
+}
+
+func anyString(values ...string) bool {
+	for _, s := range values {
+		if s != "" {
+			return true
+		}
+	}
+
+	return false
+}
 
 // Defaults applies default settings to a cobra.Command to improve the help output.
 func Defaults(c *cobra.Command) *cobra.Command {
@@ -22,4 +50,30 @@ func Defaults(c *cobra.Command) *cobra.Command {
 	c.DisableFlagsInUseLine = true
 
 	return c
+}
+
+// DefaultServer builds a gRPC server from the given flags.
+func DefaultServer(cmd *cobra.Command) (*grpc.Server, error) {
+	opts := []grpc.ServerOption{
+		grpc.MaxConcurrentStreams(1 << 20),
+	}
+
+	if anyString(
+		mustString(cmd.Flags().GetString("tls-cert-path")),
+		mustString(cmd.Flags().GetString("tls-key-path")),
+		mustString(cmd.Flags().GetString("tls-ca-path")),
+	) {
+		creds, err := auth.NewServerCredentials(
+			mustString(cmd.Flags().GetString("tls-cert-path")),
+			mustString(cmd.Flags().GetString("tls-key-path")),
+			mustString(cmd.Flags().GetString("tls-ca-path")),
+		)
+		if err != nil {
+			return nil, err
+		}
+
+		opts = append(opts, grpc.Creds(creds))
+	}
+
+	return grpc.NewServer(opts...), nil
 }
